+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+-   `FlatThemeOptions.isDark`
+    - Allows making light themes.
+
+### Changed
+
+-   `makeFlatTheme` spider chart web style color is now taken from grid-lines rather than ui border.
+
 ## [1.0.0] - 2023-02-06
 
 ### Added